kolab/issue 1658 - Adding a warning for empty To : field. Now if the To : field is empty it gives you KMessageBox::QuestionYesNo warning. If user presses Yes, the mail is sent anyway. But if the user selects No the focus is set on To: field. Earlier cases work like they used to before the change.
Allen, would you like this feature in branch/pim? I have tested it, its not so intrusive as well. I can port it if 
you think it would be a nice thing to add. I checked with my 3.5/kubuntu vanilla and enterprise kmail. It does send 
emails without any warning even if there is a empty To: field.

CCMAIL:winterz@kde.org

svn path=/branches/kdepim/proko2/kdepim/; revision=644062