forward port SVN commit 633861 by adridg:
If there is no template to use for an action, disable the menu item itself instead of inserting a lame 1-disabled-item submenu. (Well, insert the lame submenu anyway, but disable the menu item too.) Still waiting on a real usability answer on this issue, but the votes in #kontact go for this.


svn path=/trunk/KDE/kdepim/; revision=654614