Initial checkin of the openchange resource. This will (when/if done) allow akonadi users to work with Microsoft Exchange servers (for mail, calendar, tasks, etc).
This depends on using the svn trunk version of openchange,
and also relies on a recent version of samba4.

There is a long way to go, from a couple of hardwired paths
that I need to fix to the fact that it doesn't actually talk
to the server for anything yet.

Right now all it does is read the profile list, delete
entries from the profile list, and set a default profile.

Help with a nicer icon (I just borrowed an old one from
groupwise), UI design or anything else would be most 
welcome.



svn path=/trunk/KDE/kdepim/; revision=665927