RecordConduit: * Added support for fullsync * Added tests for fullsync * Made the tests and more important fullsync actualy work =:) * Removed the syncFields method and added   - copy( const Record* from, HHRecord* to )   - copy( const HHRecord* from, Record* to )
Record:
* Removed ugly == operators (ugly because only used for tests before)
* Removed fields method (over simplification)
* Added abstract equal( Record *) method.

Added cases for record syncing to the usecase document -> modified records
that don't have the modified flag set.

Adjusted tests to work with above mentioned changes.

This all actualy means that the most importants part of the base conduit are
as good as ready. Now I can start working on the keyring conduit and finetune
the base conduit.

svn path=/trunk/KDE/kdepim/; revision=693358