Ok, the foundation of kmobiletools' new architecture is finally ready.
Sorry I can't give a complete change log for this commit ;-), but the most important change is:
We can now define and implement so called "core services" which are the counter-part of an interface defined by an
engine and which automagically can determine at run-time if they are supported by the current engine (for example if the engine which
communicates with the mobile phone implements an address book interface, the corresponding address book service is loaded which
presents the user with the address book information).

So now it's possible to extend KMobileTools dynamically at runtime which opens new possibilities in user participation (e.g. you could
now easily create a plasma applet which presents phone information).


What we need to do from here is (TODO):
- remove and port old classes to the new architecture
- implement all available interfaces for the fake engine to be able to test all features (and at the same time refine the interface
definitions themselves...)
- complete at least the address book core service and create a sms core services (I don't want any feature regression)
- finally port the available engines to the new interface driven architecture

**What we need urgently is manpower. Porting the existing classes is not that hard and needs only basic Qt knowledge.**


svn path=/trunk/KDE/kdepim/; revision=700493