Backport SVN commit 703234 by tilladam:
Make sure the crypto state is not restored from headers if there are no headers.
Thanks to Alexis Papadopoulos for noticing, and for coming up with a patch 
for this in the first place.


Backport SVN commit 702320 by tilladam:

Make sure to store the state of the crypto actions when saving a mail to 
drafts, but only if the state of the mail itself should not be used as
indication of the state of the actions. This is governd by the option
to "never sign and encrypt when saving to drafts" in the security 
settings.

Prokde35-z item 44.
KDE BUG: 103240

svn path=/branches/KDE/3.5/kdepim/; revision=703251