- deprecated-- - tabs--
In this commit, I again broke config compatibility :(

The reason is that the unfiltered sernum list for IMAP
accounts is stored top-level in a config file, without
any group. 
Now, all readEntry calls without a group are deprecated,
which means we have to use a group in order to remove
KDE3SUPPORT. I used the group "Default".

It also seems that writing an update script is not
possible at all, as the .upd file wants to have a filename.
And the filename for the sernums here includes the account
ID, so there is no way to specifiy this files in the 
.upd file.

And I just also realized that I did the same with POP 
(already downloaded UID list)...

Anybody got a solution for this?
If not, the list of downloaded POP UIDs and the list
of unfiltered IMAP sernums will get lost during the
upgrade.

CCMAIL: kloecker@kde.org


svn path=/trunk/KDE/kdepim/; revision=707368