Backport of:
SVN commit 717675 by tilladam:

Change the default for what to select on entering a folder to "last selected" 
from "jump to new". This is less suprising, it seems, and avoids new mails being
missed since they are automatically marked as read when entering a folder. This
is perceived as "mail loss" by some users. It's also more predictable when
switching between two folders, which apparently a lot of people do.

svn path=/branches/kdepim/enterprise/kdepim/; revision=717678