Configure option for enabling/disabling accounts without changing any settings.
I found it useful for testing/debugging connections when only one or two (of many configured) accounts are important in the current context, so breakpoints are reached only when we do expect that.

Moreover, for some users, the option is more fine-grained than going globally offline.

This option could be later provided in the configuration dialog as well.

Implementation detail:
AccountManager skips disabled accounts when reading the config and leaves them untouched on writing back the config.


Reviewed by A.Winter


svn path=/trunk/KDE/kdepim/; revision=780603