port SVN commit r800168 by ervin:
Apply the RFC2047 decoding inside of normalizeAddressesAndDecodeIdn() as
advised by Ingo. Use the RFC2047 implementation of kmime for that matter
(yes, we have at least three implementations of this rfc in kdepim).

Since the RFC2047 decoding is now done in normalizeAddressesAndDecodeIdn(),
use the raw headers for the relevant addresses related fields in KMMessage.



svn path=/branches/KDE/3.5/kdepim/; revision=800206