- Return the correct return code when not all messages could be moved in KMMoveCommand - In the action scheduler, don't delete the original mail if the filtered copy couldn't   be moved back
This is a hotfix for IMAP, which will eat messages if one of the filters has online IMAP
target folder (in which case the action scheduler will be used).

The real problem is of course that the move command fails in the frist place, will have to
figure that out later.


svn path=/trunk/KDE/kdepim/; revision=816901