Windows: when another instance of unique appliaction is started, activate previous instance. This is what users expect.
Moreover, if standalone application is running, e.g. KMail, 
and user clicked on "e-mail" component of Kontact, window 
of the standalone application is activated.

pimapplication_win.cpp is now unused on Windows, 
as we switch to fixes at windbus level.

The code is also used in kdeui (will be commited soon, 
and shared later).


svn path=/branches/kdepim/enterprise4/kdepim/; revision=826794