port from e3 SVN commit 840224 by vkrause:
Avoid calling updateEvents() if nothing has changed. Since we get
updateView() calls for all kinds of reasons and usually three in a row,
we need to monitor for relevant changes in the calendar ourselves.

This reduces the number of calls to the KCal::Calendar::events(QDate)
method which is quite expensive for large calendars considerably.

svn path=/trunk/KDE/kdepim/; revision=841233