Manually merge 846008 from trunk, which itself was a e3->trunk merge by Allen (r804642 by vkrause).
That's a true KDE PIM bug: Operating on an invitation in the KMail part in Kontact triggers a calendar update in the KOrganizer part while being 
hidden causing the splitter synchronizer event filter installation to fail randomly for some splitters of the side-by-side view resulting in a 
totally messed up view after switching back to KOrganizer. Of course only when using side-by-side view exclusively, not in the tabbed mode. No idea 
how to really fix this, so as a workaround update the view again
after being shown if the above problem was detected while being hidden.

Kolab issue 2675

svn path=/branches/kdepim/enterprise4/kdepim/; revision=850484