Backport r847987 by tmcguire from trunk to the 4.1 branch:
Don't duplicate attachment handling code in the mime part tree.
This fixes crashes when viewing loaded-on-demand attachments via the mime tree,
since it would not set mAtmUpdate correctly, which prevents re-creating the part
tree when the part is retrived. Re-creating the part-tree would make the pointers
inside the map of the load parts command invalid and therefore trigger a crash.

CCBUG: 136145



svn path=/branches/KDE/4.1/kdepim/; revision=851216