backport SVN commit 860086 by winterz:
work around what krake and I think is a Qt bug when converting hh:mm to a QTime.
we do a silly check to see if the user supplied seconds for the start/end times,
and if not, we append ":00" for them.

i.e. "10:00" becomes "10:00:00" so QTime::fromString( Qt::ISODate) doesn't fail.

svn path=/branches/KDE/4.1/kdepim/; revision=860088