Backport r879276 by tmcguire from trunk to the 4.1 branch:
When a message is already in the metadata map (either by UID or by MD5), don't
filter it, as it (hopefully) means we have already seen and therefore filtered
the message.
This fixes filters for online IMAP running multiple times for the same message,
and might even fix the problem of duplicate messages after filtering.

CCBUG: 95064


svn path=/branches/KDE/4.1/kdepim/; revision=883868