backport SVN commit 890719 by winterz:
Use valid QDateTimes in readEntry and writeEntry, even in the cases
where we really only need to read/write a QDate or QTime.
Unfortunately, QDates and QTimes are not supported by KConfig.
and then there's the issue of legacy korganizer_printingrc files.
 
we should have used QStrings.

anyhow, this should stop the console messages:
  "conversion of "-4713,1,1,18,15,53" to QDateTime failed"

svn path=/branches/KDE/4.1/kdepim/; revision=893477