backport SVN commit 907378 by winterz:
Use static_cast instead of dynamic_cast when casting from 
a generic KPluginFactory to our KAB::ContactEditorWidgetFactory.

Else the cast can fail, especially when KAddressbook is used
as a plugin itself as part of Kontact.

Fixes the bug of the [Crypto Settings] tab not showing up in the contact editor.

svn path=/branches/KDE/4.2/kdepim/; revision=907380