Backport SVN commit 909231 by vkrause from trunk:
- fix crash in client bridge setup
- allow to omit client bridge setup 


Backport SVN commit 909347 by vkrause from trunk:

Allow to optionally omit creation of client side bridges.


This disables the creation of client side bridges in 4.2 branch, as
discussed during the KDE PIM meeting last weekend. Akonadi migration
still happens for applications using Akonadi directly (such as KPilot)
and we thus also get wide-spread testing for Akonadi deployment issues.
However, the KResource based applications will continue to use the old
API instead of the compat bridges. This is due to elaborate event loop
fun with them, in conjunction with complex applications such as Kontact.
Not the fault of the bridges, but caused by the synchronous API of KResources
which when emulated with sub-eventloops has interesting side-effects
with code that is not prepared for reentrancy. This would be fixable
with unreasonable effort or - much more useful - with porting the
problematic client code to use Akonadi directly, which will happen for
KDE 4.3.
In short: if you've seen weird behaviour in Kontact, ranging from
random crahs to POP3 mail duplication (you wouldn't expect that when
changing addressbook code, would you?), this might help.

svn path=/branches/KDE/4.2/kdepim/; revision=912022