Backport r913992 by tmcguire from trunk to the 4.2 branch:
Fix the bug that causes KMail to loose receiving accounts when
KMail is exiting and wants to open the wallet, but is killed during
that time (happens during logout).

Now, the account config groups are saved after writing out the existing
accounts, so this can't happen anymore.

I also removed the feature to disable accounts, it was only a
hidden feature and caused trouble with this.

CCBUG: 169166
CCMAIL: winter@kde.org


svn path=/branches/KDE/4.2/kdepim/; revision=913998