Add a "Maintenance" tab to the folder properties dialogue.  This gives some information about the folder, its file and index locations and sizes, and the content and messages.
For mbox folders, the compaction state and a button to do it is here (still
also in the folder context menu).  A folder can be compacted (at the user's
own risk) even if it was previously marked "unsafe".

Other dangerous operations ("Rebuild Index" and "Rebuild IMAP Cache") have
been moved from the context/menubar menus to here.

Moved the "mark_all_as_read" action back to the top of the folder popup menu,
as suggested by aseigo.

Added some hooks (KMFolderDialog::setPage() and KMMainWidget::PropsPage enum)
to embed the other folder properties into this dialogue, if there can ever be
agreement on the usability of this.  So it's not a fix for bug 115611, but it's
a start...

CCBUG:115611

svn path=/trunk/KDE/kdepim/; revision=927964