Backport r929169 by tilladam from trunk to the 4.2 branch:
Revert 928247 which forward ported a fix from enterprise35 via
enterprise4 which I believe is wrong in all branches, but doesn't crash
in e35, probably. Kleo jobs auto-delete, and thus do not need to be
deleted again from KMail. The commit to e35 put the jobs in to auto_ptrs
in KMail, which in e35 probably delete the jobs before the event loop is
next reached (and thus before the job's deleteLater()) resulting in no
crash. In e4 and trunk, because of the newly added event loop wrapping
KleoJobExectuor, the deleteLater happens before the auto_ptr dtor
resulting in a crash in that.

Marc, am I missing something or does this need to be reverted in e35 and
e4 as well?

CCMAIL: mutz@kde.org

svn path=/branches/KDE/4.2/kdepim/; revision=931333