You'll need an up-to-date e4-kdelibs for this!
Introduce the CMake variable ENTERPRISE_BUILD.
Now, a config-enterprise.h is created, which you can
#include to have enterprise-specifc behavior.

This is used in KMail for the "too many recipients"
warning and the default forward type combobox, which
I don't really want in trunk. The GUI option is disabled
here.
The default too-many-recipients warning is disabled by default for
non-enterprise builds, with some cmake magic.

svn path=/branches/kdepim/enterprise4/kdepim/; revision=935217