clean up some compile warnings. in particular, I renamed KDGanttViewItem::setText() and KDGanttViewItem::text() to setItemText() and itemText() because they were masking Q3ListViewItem virtual methods of the same names but with different signatures.
I'm not sure if KDAB cares about this old version of KDGantt anymore?
There are a couple of deprecated functions in here that could also be cleaned-up.
like bitBlt().

Of course, we in kdepim should start using the new KDGantt.
There are plans to do that, but making these little fixes was easier for now.

CCMAIL: mirko@kde.org,dfaure@kde.org

svn path=/trunk/KDE/kdepim/; revision=937849