Backport r940441 by tmcguire from trunk to the 4.2 branch:
Don't produce garbage when forwarding a mail with a filter.

For this, I had to remove the code duplication in KMFilterActionForward::process()
and use KMMessage::createForward() instead.

This uncovered a bug in createForward (content type boundary was not saved) and a
small bug in the template parser.

Also, I added a KMMessage::dump() function to help with debugging mimelib 
related troubles.

Not perfect yet, the way the template parser copies the message needs to be
improved, right now, when forwarding a signed message, the signed text gets
replaced with the template text, resulting in an invalid signature ...

CCBUG: 174942


svn path=/branches/KDE/4.2/kdepim/; revision=940772