Backport r940516 by tmcguire from trunk to the 4.2 branch:
When replacing the text of a message after processing the template, don't try
to just replace the first text/plain part, this will cause trouble with signatures
and possibly many other situations.

Instead, delete the content of the old message, and set the processed
template text as the first body part.
If the old message had attachments, re-add them to the new message.


svn path=/branches/KDE/4.2/kdepim/; revision=940775