In insertTodo(), the code that prevented infinit recursion had an infinit loop.
This happened when the to-do being inserted wasn't in a loop but it's parents were, for example:
t1 is son of t2 which is son of t3 which is son of t2.

Also added the loop check to moveIfParentChanged() which also contained an infinit loop when there's
such a to-do hierarchy.

svn path=/trunk/KDE/kdepim/; revision=942112